flowrate: merging flowrate.go and monitor.go
